+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.0
+- Resolution for typed arrays and other built-in data structure types
+- More formal test for values
+- Retain the type of objects (e.g, Set, Map etc)
+- Removed `createArr`, `createObj` and `create()` (use `json()`)
+- Renamed `call()` to `exec()`
+
 ## 1.3.1
 - Fix issue when building array